Fixed signed / unsigned comparison warnings

+++ include/coin/Script.h
@@ -651,7 +651,7 @@ public:
         opcodetype opcode;
         do
         {
-            while (end() - pc >= b.size() && memcmp(&pc[0], &b[0], b.size()) == 0)
+            while ((unsigned long)(end() - pc) >= b.size() && memcmp(&pc[0], &b[0], b.size()) == 0)
                 erase(pc, pc + b.size());
         }
         while (getOp(pc, opcode));
